refactor: move the vtbackend test doubles into testing/
MockTerm and TestHelpers ship inside the production library today, and they are
not local to it: vtconformance drives MockTerm from two of its tests, and a
vtrasterizer test includes TestHelpers. A test double consumed across module
boundaries is a published thing, and every other module in the tree already says
so with a directory -- net/testing/, vthost/testing/, crispy/testing/,
coro/testing/.

test_main.cpp, the Catch2 entry point, comes with them.

What is left at the top of src/vtbackend/ is now what belongs there: CMakeLists,
Logging.hpp, and the two optional tools (bench-headless, doc-tool). Every header
that describes the terminal lives in a directory that names what it is for.
