+++ src/vtbackend/input/vi/ViCommands_test.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/MockTerm.hpp>
-#include <vtbackend/TestHelpers.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/testing/MockTerm.hpp>
+#include <vtbackend/testing/TestHelpers.hpp>
 
 #include <vtpty/PageSize.hpp>
