+++ src/vtbackend/screen/Screen_test.cpp
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/MockTerm.hpp>
-#include <vtbackend/TestHelpers.hpp>
 #include <vtbackend/core/Primitives.hpp>
 #include <vtbackend/input/InputGenerator.hpp>
 #include <vtbackend/screen/Screen.hpp>
 #include <vtbackend/screen/Viewport.hpp>
+#include <vtbackend/testing/MockTerm.hpp>
+#include <vtbackend/testing/TestHelpers.hpp>
 #include <vtbackend/vt/Charset.hpp>
 
 #include <crispy/Escape.hpp>
