+++ src/vtbackend/shell/Folding_test.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/MockTerm.hpp>
 #include <vtbackend/screen/StatusLineBuilder.hpp>
 #include <vtbackend/screen/Terminal.hpp>
 #include <vtbackend/shell/Folding.hpp>
+#include <vtbackend/testing/MockTerm.hpp>
 
 #include <crispy/Utils.hpp>
