+++ src/vtbackend/vt/KittyClipboard_test.cpp
@@ -2,7 +2,7 @@
 //
 // Unit tests for the kitty clipboard protocol (OSC 5522) and its paste-notification mode (5522).
 
-#include <vtbackend/MockTerm.hpp>
+#include <vtbackend/testing/MockTerm.hpp>
 #include <vtbackend/vt/KittyClipboard.hpp>
 
 #include <crispy/Base64.hpp>
