+++ src/vtconformance/ScreenDump_test.cpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/MockTerm.hpp>
 #include <vtbackend/core/CellFlags.hpp>
+#include <vtbackend/testing/MockTerm.hpp>
 
 #include <catch2/catch_test_macros.hpp>
